backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.4
python-dotenv==1.0.1
httpx==0.27.0
supabase==2.5.0
langchain==0.2.3
langchain-community==0.2.4
langchain-google-genai==1.0.5
sentence-transformers==3.0.1
numpy==1.26.4
scikit-learn==1.5.0
openpyxl==3.1.2
python-multipart==0.0.9
